js: converted repositories files to vue format
