+++ app/views/namespaces/show.html.slim
@@ -1,8 +1,1 @@
 <namespaces-show-page :namespace-ref="#{@namespace_serialized}" :repositories-ref="#{@repositories_serialized}" repositories-path="#{repositories_path}" webhooks-path="#{namespace_webhooks_path(@namespace)}" teams-path="#{teams_path}"></namespaces-show-page>
-
-- content_for :js_body do
-  script#js-repositories-table-tmpl type="text/x-template"
-    = render "repositories/components/table"
-
-  script#js-repository-table-row-tmpl type="text/x-template"
-    = render "repositories/components/table_row"
