--- app/views/repositories/components/_table.html.slim
@@ -1,43 +1,0 @@
-.table-responsive
-  table.table.table-striped.table-hover.no-margin :class=="{'table-sortable': sortable}"
-    colgroup
-      col class="col-60"
-    thead
-      tr
-        th @click=="sort('name')"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'name' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'name' && !sorting.asc,
-            }"
-          }
-          | Repository
-
-        th @click=="sort('namespace.name')" v-if="showNamespaces"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'namespace.name' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'namespace.name' && !sorting.asc,
-            }"
-          }
-          | Namespace
-        th @click=="sort('tags_count')"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'tags_count' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'tags_count' && !sorting.asc,
-            }"
-          }
-          | Tags
-        th @click=="sort('updated_at')"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'updated_at' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'updated_at' && !sorting.asc,
-            }"
-          }
-          | Updated at
-    tbody
-      <repository-table-row v-for="repository in filteredRepositories" :key="repository.id" :repository="repository" :repositories-path="repositoriesPath" :namespaces-path="namespacesPath" :show-namespaces="showNamespaces"></repository-table-row>
-
-  <table-pagination :total.sync="repositories.length" :current-page="currentPage" :itens-per-page.sync="limit" @update="updateCurrentPage"></table-pagination>
