--- app/views/repositories/components/_table_row.html.slim
@@ -1,11 +1,0 @@
-tr :class="scopeClass"
-  td
-    a :href="repositoryPath"
-      | {{ repository.name }}
-  td v-if="showNamespaces"
-    a :href="namespacePath"
-      | {{ repository.namespace.name }}
-  td
-    | {{ repository.tags_count }}
-  td
-    | {{ updatedAt }}
