+++ app/views/repositories/index.html.slim
@@ -1,8 +1,1 @@
 <repositories-index-page :team-repositories-ref="#{@team_repositories_serialized}" :other-repositories-ref="#{@other_repositories_serialized}" namespaces-path="#{namespaces_path}" repositories-path="#{repositories_path}"></repositories-index-page>
-
-- content_for :js_body do
-  script#js-repositories-table-tmpl type="text/x-template"
-    = render "repositories/components/table"
-
-  script#js-repository-table-row-tmpl type="text/x-template"
-    = render "repositories/components/table_row"
