--- app/views/repositories/star.js.erb
@@ -1,20 +1,0 @@
-
-<% if @repository.errors.any? %>
-  $('#float-alert p').html("<%= escape_javascript(@repository.errors.full_messages.join('<br/>')) %>");
-  $('#float-alert').fadeIn(setTimeOutAlertDelay());
-<% else %>
-
-  var element = $('#toggle_star');
-
-  if (<%= @repository.starred_by?(user) %>) {
-    element.attr('data-original-title', 'Unstar repository');
-    element.find('.fa').removeClass('fa-star-o');
-    element.find('.fa').addClass('fa-star');
-  } else {
-    element.attr('data-original-title', 'Star repository');
-    element.find('.fa').removeClass('fa-star');
-    element.find('.fa').addClass('fa-star-o');
-  }
-  $('#star-counter').text(<%= @repository.stars.count %>);
-
-<% end %>
